Use uint16 instead of uint32, speedup 10%

+++ heuristic.c
@@ -14,8 +14,8 @@
 
 /* For sorting */
 static int compare(const void *lhs, const void *rhs) {
-    int lhs_integer = *(const int *)(lhs);
-    int rhs_integer = *(const int *)(rhs);
+    int16_t lhs_integer = *(const int16_t *)(lhs);
+    int16_t rhs_integer = *(const int16_t *)(rhs);
     return rhs_integer - lhs_integer;
 }
 
@@ -24,7 +24,7 @@ static int compare(const void *lhs, const void *rhs) {
  * Charset size over sample
  * will be small <= 64
  */
-static int _symbset_calc(const uint32_t *bucket)
+static int _symbset_calc(const uint16_t *bucket)
 {
     uint32_t a;
     uint32_t symbset_size = 0;
@@ -43,7 +43,7 @@ static int _symbset_calc(const uint32_t *bucket)
  * > 200 - bad compressible data
  * For right & fast calculation bucket must be reverse sorted
  */
-static int _coreset_calc(const uint32_t *bucket,
+static int _coreset_calc(const uint16_t *bucket,
     const uint32_t sum_threshold)
 {
     uint32_t a = 0;
@@ -58,7 +58,7 @@ static int _coreset_calc(const uint32_t *bucket,
     return a;
 }
 
-static int _entropy_perc(const uint32_t *bucket, const uint32_t sample_size)
+static int _entropy_perc(const uint16_t *bucket, const uint32_t sample_size)
 {
     uint32_t a, p;
     uint32_t entropy_sum = 0;
@@ -108,7 +108,7 @@ enum compress_advice heuristic(const uint8_t *input_data,
     const uint32_t shift = bytes_len/offset_count;
     const uint32_t sample_size = offset_count*READ_SIZE;
     uint32_t a, b;
-    uint32_t bucket[256];
+    uint16_t bucket[256];
     uint8_t  *sample = malloc(sample_size);
 
     for (a = 0; a < BUCKET_SIZE; a++)
@@ -135,7 +135,7 @@ enum compress_advice heuristic(const uint8_t *input_data,
     }
 
     /* Sort in reverse order */
-    sort(bucket, BUCKET_SIZE, sizeof(uint32_t), &compare, NULL);
+    sort(bucket, BUCKET_SIZE, sizeof(uint16_t), &compare, NULL);
 
     a = _coreset_calc(bucket, sample_size*90/100);
